Fix #2. Rewrite with tests.
